gta4xl-common: Encryption upgrades
This:
 - Upgrades the data partition to use FBE v2.
 - Enables metadata encryption, with the newer dm-default-key API.
 - Sets inlinecrypt for /data. Samsung's SM-E625F stock firmwares have
   it set, while the Exynos FMP kernel driver and other bits have been
   taken from some E625F kernel sources, and all at least appears
   unharmed with it.
 - Sets properties for enabling FBE v2 and use of dm-default-key on
   older shipping API levels as per [1].

 1: https://source.android.com/docs/security/features/encryption/metadata

With these upgrades and the switch to F2FS, adoptable storage appears to
be working fine.

Lastly, this also adds a separate recovery fstab without the flag for
enabling metadata encryption set on data. On other devices, that has
been observed to cause recovery to not format the data partition with a
proper filesystem, and then issues on booting after a factory reset.

Apparently, a property, ro.crypto.metadata_init_delete_all_keys.enabled,
can be set to true for some effect/a fix. But, on another device, I
observed it to seemingly have no effective result on getting rid of the
issues on boot. I haven't tested it here, and neither can I verify it
working while being unable to do another factory reset.

That said, a bulk of other device trees also have this separate fstab
setup, and that property gets little use.

So, keep the separate fstab to be absolutely safe.

Change-Id: I4e46e4d6b023b15eaa4bc1b6feb15b4b4f97a052